+++ OpenBudgeteer.Core.Data.Postgres.Migrations/20231221153511_AddBucketTextColor.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace OpenBudgeteer.Core.Data.Postgres.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddBucketTextColor : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "TextColorCode",
+                table: "Bucket",
+                type: "text",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "TextColorCode",
+                table: "Bucket");
+        }
+    }
+}