+++ OpenBudgeteer.Core.Data.Postgres.Migrations/DatabaseContextModelSnapshot.cs
@@ -17,7 +17,7 @@ namespace OpenBudgeteer.Data.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "7.0.5")
+                .HasAnnotation("ProductVersion", "8.0.0")
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
@@ -89,6 +89,9 @@ namespace OpenBudgeteer.Data.Migrations
                         .HasColumnType("timestamp without time zone");
 
                     b.Property<string>("Name")
+                        .HasColumnType("text");
+
+                    b.Property<string>("TextColorCode")
                         .HasColumnType("text");
 
                     b.Property<DateTime>("ValidFrom")
@@ -369,7 +372,7 @@ namespace OpenBudgeteer.Data.Migrations
             modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.Bucket", b =>
                 {
                     b.HasOne("OpenBudgeteer.Core.Data.Entities.Models.BucketGroup", "BucketGroup")
-                        .WithMany()
+                        .WithMany("Buckets")
                         .HasForeignKey("BucketGroupId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -380,7 +383,7 @@ namespace OpenBudgeteer.Data.Migrations
             modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.BucketMovement", b =>
                 {
                     b.HasOne("OpenBudgeteer.Core.Data.Entities.Models.Bucket", "Bucket")
-                        .WithMany()
+                        .WithMany("BucketMovements")
                         .HasForeignKey("BucketId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -402,7 +405,7 @@ namespace OpenBudgeteer.Data.Migrations
             modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.BucketVersion", b =>
                 {
                     b.HasOne("OpenBudgeteer.Core.Data.Entities.Models.Bucket", "Bucket")
-                        .WithMany()
+                        .WithMany("BucketVersions")
                         .HasForeignKey("BucketId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -413,13 +416,13 @@ namespace OpenBudgeteer.Data.Migrations
             modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.BudgetedTransaction", b =>
                 {
                     b.HasOne("OpenBudgeteer.Core.Data.Entities.Models.Bucket", "Bucket")
-                        .WithMany()
+                        .WithMany("BudgetedTransactions")
                         .HasForeignKey("BucketId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.HasOne("OpenBudgeteer.Core.Data.Entities.Models.BankTransaction", "Transaction")
-                        .WithMany()
+                        .WithMany("BudgetedTransactions")
                         .HasForeignKey("TransactionId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -443,7 +446,7 @@ namespace OpenBudgeteer.Data.Migrations
             modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.MappingRule", b =>
                 {
                     b.HasOne("OpenBudgeteer.Core.Data.Entities.Models.BucketRuleSet", "BucketRuleSet")
-                        .WithMany()
+                        .WithMany("MappingRules")
                         .HasForeignKey("BucketRuleSetId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -460,6 +463,30 @@ namespace OpenBudgeteer.Data.Migrations
                         .IsRequired();
 
                     b.Navigation("Account");
+                });
+
+            modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.BankTransaction", b =>
+                {
+                    b.Navigation("BudgetedTransactions");
+                });
+
+            modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.Bucket", b =>
+                {
+                    b.Navigation("BucketMovements");
+
+                    b.Navigation("BucketVersions");
+
+                    b.Navigation("BudgetedTransactions");
+                });
+
+            modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.BucketGroup", b =>
+                {
+                    b.Navigation("Buckets");
+                });
+
+            modelBuilder.Entity("OpenBudgeteer.Core.Data.Entities.Models.BucketRuleSet", b =>
+                {
+                    b.Navigation("MappingRules");
                 });
 #pragma warning restore 612, 618
         }